# Timing info in external test JSON reports