Gathering info about compilation time in external tests to create a comparison like https://github.com/ethereum/solidity/pull/14909#issuecomment-1981534920 is incredibly time consuming. I have to navigate to 10-15 CI pages, locate `time` output in the long CI log for each, copy it and manually format it info something human-readable, like a table. This PRs is the first step toward automating this somewhat.

Now that info will be present in the combined JSON report from all external tests and I'll be able to pull it out with a simple script. This is just the bare minimum to make it less annoying for me. The extra data is not yet processed by the scripts that format gas tables or diff them. For now I'm planning to create a quick, hacky script to do further processing but if it's reusable enough I might submit it in a follow-up PR.

The new info can be found in the `reports/externalTests/all-benchmarks.json` artifact of the `c_ext_benchmarks` job, under the `<project>.<preset>.compilation_time` keys. It's also present in the same form in individual reports attached as artifacts to each parallel run of each external test job.