@cameel 2024-04-15
So, with this PR merged, this is the hacky way to quickly get a timing table from external test benchmarks:

### Script
```bash
jq '[
    ["brink",            ."brink"            ."ir-optimize-evm+yul".compilation_time.user],
    ["colony",           ."colony"           ."ir-optimize-evm+yul".compilation_time.user],
    ["elementfi",        ."elementfi"        ."ir-optimize-evm+yul".compilation_time.user],
    ["ens",              ."ens"              ."ir-optimize-evm+yul".compilation_time.user],
    ["euler",            ."euler"            ."ir-optimize-evm+yul".compilation_time.user],
    ["gnosis",           ."gnosis"           ."ir-optimize-evm+yul".compilation_time.user],
    ["gp2",              ."gp2"              ."ir-optimize-evm+yul".compilation_time.user],
    ["perpetual-pools",  ."perpetual-pools"  ."ir-optimize-evm+yul".compilation_time.user],
    ["pool-together",    ."pool-together"    ."ir-optimize-evm+yul".compilation_time.user],
    ["uniswap",          ."uniswap"          ."ir-optimize-evm+yul".compilation_time.user],
    ["yield_liquidator", ."yield_liquidator" ."ir-optimize-evm+yul".compilation_time.user],
    ["zeppelin",         ."zeppelin"         ."ir-optimize-evm+yul".compilation_time.user]
]' all-benchmarks.json | python -c "$(cat <<EOF
import tabulate, json, sys
print(tabulate.tabulate(json.load(sys.stdin), tablefmt="github", headers=["Test", "Time"]))
EOF
)"
```

Could be improved to avoid having to list the keys by hand, and to round the numbers, but for now it's good enough.

### Result
| Test             |    Time |
|------------------|---------|
| brink            |   5.969 |
| colony           | 110.336 |
| elementfi        | 168.006 |
| ens              |  41.883 |
| euler            |  64.441 |
| gnosis           |         |
| gp2              |  40.769 |
| perpetual-pools  |  77.294 |
| pool-together    |  83.363 |
| uniswap          | 105.879 |
| yield_liquidator |  30.584 |
| zeppelin         | 263.092 |