@cameel 2024-04-19
Here are some more streamlined scripts, with automatic iteration over keys, rounding and downloading of benchmark results. To use them, just fill in the branch name and paste it into shell.

### Timing of a single branch
```bash
branch="<BRANCH NAME HERE>"
preset=ir-optimize-evm+yul

function timing-table-script {
    cat <<EOF
import tabulate, json, sys

def as_seconds(value):
    return (str(value) + " s") if value is not None else None

table = json.load(sys.stdin)
table = [[row[0], as_seconds(row[1])] + row[2:] for row in table]

headers = ["Project", "Time"]
alignment = ("left", "right")
print(tabulate.tabulate(table, tablefmt="pipe", headers=headers, colalign=alignment))
EOF
}

function tabulate-ext-timing {
    python -c "$(timing-table-script)" "$@"
}

function ext-timing-list {
    local preset="$1"
    jq "[to_entries[] | [.key, (.value.\"${preset}\".compilation_time.user | if . != null then round else . end)]]" "${@:2}"
}

scripts/externalTests/download_benchmarks.py --branch "$branch"
cat "all-benchmarks-${branch}"*.json | ext-timing-list "$preset" | tabulate-ext-timing
```

| Project          |   Time |
|:-----------------|-------:|
| brink            |    5 s |
| colony           |  105 s |
| elementfi        |  172 s |
| ens              |   40 s |
| euler            |   52 s |
| gnosis           |        |
| gp2              |   42 s |
| perpetual-pools  |   81 s |
| pool-together    |   56 s |
| uniswap          |   82 s |
| yield_liquidator |   26 s |
| zeppelin         |  261 s |

### Timing comparison between two branches
```bash
before_branch=develop
after_branch="<BRANCH NAME HERE>"
preset=ir-optimize-evm+yul

function diff-table-script {
    cat <<EOF
import tabulate, json, sys

def time_diff(before, after):
    return (after - before) if after is not None and before is not None else None
def as_seconds(value):
    return (str(value) + " s") if value is not None else None

data = json.load(sys.stdin)
table = [[
    project,
    as_seconds(data[0][project]),
    as_seconds(data[1][project]),
    as_seconds(time_diff(data[1][project], data[0][project])),
] for project in data[0].keys()]

headers = ["Project", "Before", "After", "Diff"]
alignment = ("left", "right", "right", "right")
print(tabulate.tabulate(table, tablefmt="pipe", headers=headers, colalign=alignment))
EOF
}

function tabulate-ext-timing-diff {
    python -c "$(diff-table-script)" "$@"
}

function ext-timing-dict {
    local preset="$1"
    jq "[to_entries[] | {(.key): (.value.\"${preset}\".compilation_time.user | if . != null then round else . end)}] | add" "${@:2}"
}

scripts/externalTests/download_benchmarks.py --branch "$after_branch"
scripts/externalTests/download_benchmarks.py --branch "$before_branch"
{
    cat "all-benchmarks-${before_branch}"*.json | ext-timing-dict "$preset"
    cat "all-benchmarks-${after_branch}"*.json  | ext-timing-dict "$preset"
} | jq --slurp | tabulate-ext-timing-diff
```

| Project          |   Before |   After |   Diff |
|:-----------------|---------:|--------:|-------:|
| brink            |      5 s |     4 s |    1 s |
| colony           |    105 s |   113 s |   -8 s |
| elementfi        |    172 s |   162 s |   10 s |
| ens              |     40 s |    43 s |   -3 s |
| euler            |     52 s |    59 s |   -7 s |
| gnosis           |          |         |        |
| gp2              |     42 s |    48 s |   -6 s |
| perpetual-pools  |     81 s |    72 s |    9 s |
| pool-together    |     56 s |    49 s |    7 s |
| uniswap          |     82 s |    84 s |   -2 s |
| yield_liquidator |     26 s |    27 s |   -1 s |
| zeppelin         |    261 s |   266 s |   -5 s |